Post-review: share more common code.
- Both lima and WSL need to try to get an updated kubectl via kuberlr
- Create a window helper for showing a message box
